Added option to read whole file
Added option to read whole file if caller send -1
as the buflen to read function.

Also added a copy of getsize to File class, now called
fgetsize, which returns the size of the file as
reported by fstat

Removed flags paramater from read function as it is never
used by libgfapi.

Change-Id: Ia1c04ef9717d9ca098e1961300d6ee216381c989
Signed-off-by: Thiago da Silva <thiago@redhat.com>
Signed-off-by: Prashanth Pai <ppai@redhat.com>