Fix reading of binary data in read()
As per the current code, this is the behavior:

>>> with v.fopen("/abc", 'r') as f:
...     data = f.read(5)
>>> print data
<ctypes.c_char_Array_2 object at 0x7fda7d6bbb90>
>>> print data.value
hello
>>>

It's incorrect to return a ctypes internal object back to the user.

In Python 2.x, read() always returns a string. It's really upto
the consumer to decode this string into whatever encoding it was
written with.

This patch reverts parts of this old change:
Ia2bb47343880cbf7121fed9510e4cfa085fe23bd

Change-Id: Ia1d3e5834be2b856776bd3cf8382a17ffd61d5df
Signed-off-by: Prashanth Pai <ppai@redhat.com>