Raise Exception if ctypes.CDLL() cannot load .so
Usually, this failure happens in source installs of glusterfs.
find_library() is able to find the .so file but ctypes.CDLL()
cannot load it!
Refer: http://bugs.python.org/issue18502

With this change, users are advised to modify LD_LIBRARY_PATH
when exception is raised

Also, passed the LD_LIBRARY_PATH env variable to tox environment
so that unit tests and functional tests can run.

Change-Id: Iffc5633088b3886739a8534692db88db7c3d02b7
Signed-off-by: Prashanth Pai <ppai@redhat.com>