Fix parsing of binary data as str
This change will return the raw string read as is which fixes
case where arbitrary binary data (such as an array.array) is written
and read back as string to be unmarshalled back as array.array

This patch does not yet address writing array.array and arbitrary
objects into binary files. That will be sent as a separate change.

Change-Id: Ic135198cca2c0dc8e103176920056e9df0c00f2b
Signed-off-by: Prashanth Pai <ppai@redhat.com>